Make the wait interval a variable.
When the script starts the gnome-terminal, it is expected to be
the first instance of the terminal, and it will usually take some
time for it to come up.  The time the script has to wait is now
a variable which can be modified depending on the machine (or
load) the script is running on.