Make the terminal command a variable.
Initially this screen was designed to fit gnome-terminal.
However, as this dependency is not at all required, it made sense
to remove it.  This is done in this commit: the user can modify
the name of the terminal command and thus use this script under
any EWMH/NetWM compatible window manager.