Bring the terminal to top when it is not the topmost window.
Usually the script would just toggle the visibility of the window.
However, since having the terminal window always on top is
annoying, the script does not set the always-on-top flag of the
window.  When it is invoked while the window is visible, it
first checks whether it is on top.  It if is not, it just brings
it there.  The window is hidden only if it is already the topmost
window.