Decide what action to take basing on whether the terminal has focus.
The idea behind the whole script is to make a terminal window
stand ready at hand.  This essentially means that the user want
to get into the terminal when they are not there and want to hide
the terminal when they are already there.  Now the script gets
the active window and it is not its managed window, it put focus
into its window.  If the focus is already in the terminal, it
hides the terminal.

This new strategy removes the necessity of keeping the current
window state in the control file which not only simplifies the
script itself, but also considerably reduces the number of I/O
operations performed (1 write at each initialization and 1 read
at each subsequent invocation).